first try: refactoring the entire programme to use sqlite db as the base instead of sheets on disk
