+++ .gitignore
@@ -10,7 +10,7 @@ compare/*.*
 *.parquet
 
 *.html
-*.sqlite3
+*.sqlite3*
 easy_access/classification/api_keys.py
 api_keys.py
