+++ easy_access/sheets/enrichment.py
@@ -1,402 +1,14 @@
 from easy_access.utils import info, warn, print, determine_course_code
-from easy_access.settings import SETTINGS, FileSetting, OSIRIS_DATA
+from easy_access.settings import SETTINGS, FileSetting
 import polars as pl
 import json
 import bs4
 import re
 import httpx
 import asyncio
-from dataclasses import dataclass, field
 from loguru import logger
-from nameparser import HumanName
 import Levenshtein
-from easy_access.orm.db import load_base_data
-
-# Dataclasses for osiris_contact parsing / matching
-@dataclass(frozen=True)
-class Faculty:
-    abbreviation: str = field(default="", compare=True)
-    name: str = field(default="", compare=True)
-
-    def __str__(self):
-        return self.abbreviation
-
-DEFAULT_FACULTIES = {
-        'EEMCS': Faculty(abbreviation='EEMCS', name='Electrical Engineering, Mathematics and Computer Science'),
-        'BMS': Faculty(abbreviation='BMS', name='Behavioural, Management and Social Sciences'),
-        'TNW': Faculty(abbreviation='TNW', name='Science and Technology'),
-        'ET': Faculty(abbreviation='ET', name='Engineering Technology'),
-        'ITC': Faculty(abbreviation='ITC', name='ITC Faculty'),
-    }
-
-@dataclass(frozen=True)
-class Department:
-    abbreviation: str = ""
-    name: str = ""
-    faculty: Faculty = None
-
-    def __str__(self):
-        if self.faculty:
-            return f"{self.faculty.abbreviation}-{self.abbreviation}"
-        else:
-            return self.abbreviation
-
-@dataclass(frozen=True)
-class Group:
-    abbreviation: str = ""
-    name: str = ""
-    department: Department = None
-
-    def __str__(self):
-        if self.department:
-            if self.department.faculty:
-                return f"{self.department.faculty.abbreviation}-{self.department.abbreviation}-{self.abbreviation}"
-            else:
-                return f"{self.department.abbreviation}-{self.abbreviation}"
-        else:
-            return self.abbreviation
-
-@dataclass
-class Contact:
-    raw_input_data: list[dict[str,str]] = field(default_factory=list, compare=False)
-    name: str = ""
-    email: str = ""
-    groups: set[Group] = field(default_factory=set)
-    departments: dict[str, Department] = field(default_factory=dict)
-    faculties: dict[str, Faculty] = field(default_factory=dict)
-
-    def parse_raw_input(self) -> None:
-        # parse all raw_input into Group/Department/Faculty instances
-        # and assign them to the correct list(s)
-
-        def add_faculty(abbr: str, name: str = "") -> Faculty | None:
-            try:
-                if abbr in self.faculties:
-                    return self.faculties[abbr]
-                if abbr in DEFAULT_FACULTIES:
-                    faculty = DEFAULT_FACULTIES[abbr]
-                elif abbr == "Department":
-                    faculty = None
-                else:
-                    faculty = Faculty(abbreviation=abbr, name=name)
-                self.faculties[abbr] = faculty
-                return faculty
-            except Exception as e:
-                warn(f"Error parsing faculty from raw input {abbr}, {name}: {e}")
-                return None
-
-        def add_department(abbr: str, name: str, faculty: Faculty | None, full_abbr: str) -> Department | None:
-            try:
-                if full_abbr in self.departments:
-                    return self.departments[full_abbr]
-                department = Department(abbreviation=abbr, name=name, faculty=faculty)
-                self.departments[full_abbr] = department
-                return department
-            except Exception as e:
-                warn(f"Error parsing department from raw input {abbr}, {name}, {faculty}: {e}")
-                return None
-
-        def add_group(abbr: str, name: str, department: Department) -> Group | None:
-            try:
-                group = Group(abbreviation=abbr, name=name, department=department)
-                self.groups.add(group)
-                return group
-            except Exception as e:
-                warn(f"Error parsing group from raw input {raw_input}: {e}")
-                return None
-
-        for raw_input in self.raw_input_data:
-            abbr = raw_input.get("abbr", "")
-            name = raw_input.get("name", "")
-
-            if '-' not in abbr:
-                # should be a faculty
-                add_faculty(abbr, name)
-            elif abbr.count('-') == 1:
-                # abbr should be FACULTYABBR - DEPARTMENTABBR
-                # name should be department name
-                faculty_abbr, dept_abbr = abbr.split('-')
-                dept_name = name
-                faculty = add_faculty(faculty_abbr)
-                add_department(dept_abbr, dept_name, faculty, abbr)
-            elif abbr.count('-') == 2:
-                # abbr should be FACULTYABBR - DEPARTMENTABBR - GROUPABBR
-                # name should be group name
-                faculty_abbr, dept_abbr, group_abbr = abbr.split('-')
-                full_dept_abbr = f"{faculty_abbr}-{dept_abbr}"
-                group_name = name
-                faculty = add_faculty(faculty_abbr)
-                department = add_department(dept_abbr, "", faculty, full_dept_abbr)
-                add_group(group_abbr, group_name, department)
-
-    def add_raw_input(self, raw_input: dict[str,str]) -> None:
-        if raw_input not in self.raw_input_data:
-            self.raw_input_data.append(raw_input)
-
-    def update(self) -> tuple[list[Group],list[Department],list[Faculty]]:
-        for group in self.groups:
-            if group.department:
-                if str(group.department) not in self.departments:
-                    self.departments[str(group.department)] = group.department
-
-        for dept in self.departments.values():
-            if dept.faculty:
-                if str(dept.faculty) not in self.faculties:
-                    self.faculties[str(dept.faculty)]=dept.faculty
-
-    def get_data(self) -> dict[str,str|list[Group]|list[Department]|list[Faculty]]:
-        """
-        Returns a dict with sheet colnames as keys and deduplicated data as values
-        """
-        self.parse_raw_input()
-        self.update()
-
-        return {
-            "contact_name": self.name,
-            "contact_email": self.email,
-            "contact_groups": list(self.groups),
-            "contact_departments": list(self.departments.values()),
-            "contact_faculties": list(self.faculties.values()),
-            }
-
-@dataclass
-class ItemContacts:
-    """
-    Keeps track of all Contacts for a single item
-    Mainly used to output strs with found unique depts/groups/faculties
-    """
-    contacts: list[Contact] = field(default_factory=list)
-
-    def get_merged_contact_data(self) -> dict[str,str]:
-        """
-        Returns a dict with sheet colnames as keys with merged unique strings as values
-        """
-        data: dict[str, set[str|Group|Department|Faculty]] = {
-            "contact_name": set(),
-            "contact_email": set(),
-            "contact_groups": set(),
-            "contact_departments": set(),
-            "contact_faculties": set()
-        }
-        for contact in self.contacts:
-            contact_data = contact.get_data()
-            for key, value in contact_data.items():
-                if isinstance(value, list):
-                    data[key].update(value)
-                else:
-                    data[key].add(value)
-
-        final_data: dict[str,str] = dict()
-        for key, value in data.items():
-            if not value:
-                value = ""
-            elif len(value) == 1:
-                value = str(value.pop())
-            else:
-                value = " | ".join([str(x) for x in list(value) if x])
-            final_data[key] = value
-
-        return final_data
-
-
-def enrich_df_with_osiris_data(df: pl.DataFrame, group:str = "all items") -> pl.DataFrame:
-    """
-    Read in OSIRIS/people page data.
-    Enrich the supplied df with the information contained in the jsons.
-    Return the enriched dataframe.
-    """
-
-    item_data = df.select(
-        pl.col("course_code"), pl.col("course_name"), pl.col("material_id")
-    ).to_dicts()
-    enriched_item_data = []
-    osiris_cat_link = "https://utwente.osiris-student.nl/onderwijscatalogus/extern/cursus/zoek?trefwoord="
-    info(f"Enriching {len(item_data)} items for {group} with OSIRIS data.")
-    total = len(item_data)
-    updated = 0
-    already_enriched = 0
-    no_course_code_found = set()
-    no_osiris_data_found = set()
-    for item in item_data:
-        already_found_codes = []
-        if item.get('osiris_course_codes_found'):
-            if isinstance(item['osiris_course_codes_found'], str):
-                if ' | ' in item['osiris_course_codes_found']:
-                    already_found_codes = item['osiris_course_codes_found'].split(" | ")
-                else:
-                    already_found_codes = already_found_codes.append(item['osiris_course_codes_found'])
-                already_found_codes = [i.strip() for i in already_found_codes]
-
-        if already_found_codes:
-            already_enriched += 1
-            continue
-
-        course_codes = determine_course_code(
-            item["course_code"], item["course_name"]
-        )
-
-        if not course_codes:
-            no_course_code_found.add(item["course_code"])
-            continue
-
-        course_codes = list(course_codes)
-
-
-        course_codes = [i.strip() for i in course_codes]
-
-
-        new_item = dict()
-        new_item["material_id"] = item["material_id"]
-        if len(course_codes) == 1:
-            new_item["osiris_course_codes_found"] = course_codes[0]
-        if len(course_codes) > 1:
-            new_item["osiris_course_codes_found"] = " | ".join(course_codes)
-        used_codes = []
-        found_osiris_data = []
-        maybe_used_codes = []
-        maybe_found_data = []
-        for code in course_codes:
-            new_osiris_data = OSIRIS_DATA.get(code, None) if OSIRIS_DATA else None
-            if not new_osiris_data:
-                no_osiris_data_found.add(code)
-            else:
-                if new_osiris_data.get('faculty','') == item.get('faculty'):
-                    used_codes.append(code)
-                    found_osiris_data.append(new_osiris_data)
-                else:
-                    maybe_used_codes.append(code)
-                    maybe_found_data.append(new_osiris_data)
-
-        if not found_osiris_data:
-            if maybe_found_data and maybe_used_codes:
-                used_codes = maybe_used_codes
-                found_osiris_data = maybe_found_data
-
-        if len(used_codes) == 1:
-            used_code = used_codes[0]
-            new_item["osiris_course_codes_used"] = used_code
-            new_item["osiris_course_code_data_selected"] = used_code
-
-            new_item["osiris_catalogue_url"] = osiris_cat_link + used_code
-            new_item["osiris_programme"] = found_osiris_data[0].get("programme")
-
-        elif len(used_codes) > 1:
-            new_item["osiris_course_codes_used"] = " | ".join(used_codes)
-            new_item["osiris_course_code_data_selected"] = used_codes[0]
-            new_item["osiris_catalogue_url"] = osiris_cat_link + used_codes[0]
-            new_item["osiris_programme"] = " | ".join([osiris_data.get("programme", "") for osiris_data in found_osiris_data if osiris_data.get("programme")])
-            used_code = " | ".join(used_codes)
-
-        course_contacts = ItemContacts()
-        for osiris_data in found_osiris_data:
-            if osiris_data.get("contacts"):
-                contacts: dict[str,dict[str, str|list[dict[str,str]]]] = osiris_data.get("contacts")
-                for contact_name, contact in contacts.items():
-                    faculty_dict_start = {}
-                    if contact.get('faculty'):
-                        if contact.get('faculty') in DEFAULT_FACULTIES:
-                            faculty_dict_start = {contact.get("faculty"): DEFAULT_FACULTIES.get(contact.get("faculty"))}
-                    cur_contact = Contact(name=contact_name, email=contact.get("email"), faculties=faculty_dict_start)
-                    if contact.get('orgs'):
-                        [cur_contact.add_raw_input(org) for org in contact.get('orgs')]
-                    course_contacts.contacts.append(cur_contact)
-
-        new_item.update(course_contacts.get_merged_contact_data())
-        updated += 1
-        enriched_item_data.append(new_item)
-
-    enriched_items_df = pl.DataFrame(enriched_item_data)
-    df = df.join(enriched_items_df, on="material_id", how="left")
-
-    for col in df.columns:
-        # TODO: improve handling of conflicting values
-
-        if col.endswith("_left") or col.endswith("_right"):
-            base = col.replace("_left","").replace("_right","")
-            # Drop if suffix column is all null
-            if df.select(pl.col(col).is_null().all()).item(0,0):
-                df = df.drop(col)
-                continue
-            # If base column exists, compare
-            if base in df.columns:
-
-                same_vals = df.select((pl.col(base).fill_null(value='') == pl.col(col).fill_null(value='')).all()).item(0,0)
-                if same_vals:
-                    df = df.drop(col)
-                else:
-
-                    # If base is all null, replace it
-                    if df.select(pl.col(base).is_null().all()).item(0,0):
-                        df = df.drop(base)
-                        df = df.rename({col: base})
-                    # if not: we have 2 cols that do not match. Merge them in some way.
-                    else:
-                        final_data = []
-                        #warn(f"Conflicting values for column {base} in {group} enrichment results. Overwriting with new values.")
-                        all_base_vals = df.select(pl.col(base)).to_series().to_list()
-                        all_new_vals = df.select(pl.col(col)).to_series().to_list()
-                        material_ids = df.select(pl.col("material_id")).to_series().to_list()
-                        mismatched_cols = set()
-                        if base == 'osiris_catalogue_url':
-                            #info('Not printing comparison for osiris_catalogue_urls.')
-                            ...
-                        elif len(all_base_vals) != len(all_new_vals):
-                            #warn(f'Cannot print comparison: cols are not the same length. \nlen({base})={len(all_base_vals)} != len({col})={len(all_new_vals)}')
-                            ...
-                        else:
-                            #info("Mismatches found:")
-                            printstr =""
-                            for material_id, orig_val, new_val in zip(material_ids, all_base_vals, all_new_vals):
-                                if not orig_val:
-                                    orig_val = ""
-                                if not new_val:
-                                    new_val = ""
-                                orig_data = orig_val
-                                new_data = new_val
-
-                                if ' | ' in orig_val:
-                                    orig_val = orig_val.split(' | ')
-                                    if isinstance(orig_val,list):
-                                        if len(orig_val) > 1:
-                                            orig_val = set(orig_val)
-                                if ' | ' in new_val:
-                                    new_val = new_val.split(' | ')
-                                    if isinstance(new_val,list):
-                                        if len(new_val) > 1:
-                                            new_val = set(new_val)
-
-                                if orig_val != new_val:
-                                    if not orig_val or orig_val == "":
-                                        final_data.append({'material_id':material_id, base:new_data})
-                                        printstr += (f'\nmaterial_id: {material_id}\n  {str(orig_val)}\n    --> {str(new_val)}')
-                                    elif not new_val or new_val == "":
-                                        final_data.append({'material_id':material_id, base:orig_data})
-                                        printstr += (f'\nmaterial_id: {material_id}\n  Keeping  {str(orig_val)}\n  NOT replacing with {str(new_val)}')
-                                    else:
-                                        final_data.append({'material_id':material_id, base:new_data})
-                                        printstr += (f'\nmaterial_id: {material_id}\n  Keeping  {str(orig_val)}\n  NOT replacing with {str(new_val)}')
-                                    mismatched_cols.add(material_id)
-                                else:
-                                    # no new val
-                                    final_data.append({'material_id':material_id, base:orig_data})
-                            #info(printstr)
-
-                        if not final_data:
-                            df = df.drop(base).rename({col: base})
-                        else:
-                            # drop base and col from orig df
-                            # create temp dataframe from final_data -- has cols material_id and base
-                            # join temp dataframe to df on material_id
-                            #info(f'replacing {base} with col constructed from {len(final_data)} items')
-                            tmp_df = pl.from_dicts(final_data, strict=False, infer_schema_length=1000)
-                            #info(f'tmp_df details: {tmp_df.shape[0]} rows, cols: {tmp_df.columns}')
-                            #info(tmp_df)
-                            df = df.drop([base, col]).join(tmp_df, on="material_id", how="left")
-            else:
-                # base doesn't exist? weird, just rename suffix column to base and done
-                df = df.rename({col: base})
-    #info(f"{group} enrichment results\n----------------------------\nUpdated:          {updated}/{total}\nAlready enriched: {already_enriched}/{total}\nNot in osiris:    {len(no_osiris_data_found)}/{total}\nNo coursecode:    {len(no_course_code_found)}/{total}")
-    return df
+from easy_access.db.ingest import load_base_data
 
 async def update_osiris_data(df: pl.DataFrame, only_retrieve_missing: bool = False) -> None:
     """
